Update protobuf operations to match PSA crypto 1.0.0
This commit improves on the protobuf operations we expose to the
interface, bringing it up to speed with the PSA crypto 1.0.0 spec.
The changes also include removing the key lifetime as a parameter, as
this is expected to be chosen through the choice of backend (which will
each support exactly one key lifetime/one type of key storage).
This commit also adds the following changes:
* creates an algorithm message in its own file implementing the
psa_algorithm_t type from the PSA Crypto API
* encodes most things as message and use oneof extensively
* rename OpXXX and ResultXXX to Operation and Result, the XXX is given
by the file's namespace
* modify some field names to match the API
* remove reserved fields

Co-authored-by: Hugues de Valon <hugues.devalon@arm.com>
Signed-off-by: Ionut Mihalcea <ionut.mihalcea@arm.com>