Add provider ID in list_opcodes operation
This commit adds a provider ID field in the `list_opcodes` operation
contract; it is needed as the provider of interest (i.e. for which the
opcodes should be retrieved) cannot be assumed to be the one mentioned
in the request header. That is because the request itself needs to
always be directed to the CoreProvider.

Signed-off-by: Ionut Mihalcea <ionut.mihalcea@arm.com>